Removing app teardown on login host change.
Turns out, when there's no need to specify a new window/view setup with the implementation of the OAuth plugin, there's also no need to tear down the app and start over.  This was actually causing some crash problems with login host switching, which we should look into in advance of making that functionality public in PhoneGap.

When logout or login host changes are detected now, the app simply clears the old auth credentials and redirects to the bootstrap page.