Taking unstripped hybrid framework.
There seems to be an issue with the SalesforceHybridSDK framework, where using the stripped framework causes crashes when going in and out of Airplane Mode, in conjunction with backgrounding and foregrounding the app.  The unstripped version doesn't do this, so we'll use that until we can track down the root cause of the issue.