Adding (back) CordovaResources.bundle.
This is essentially all of the resources that a Cordova application will have, wrapped into one bundle.  Some further thought needs to be put into how this information is managed, since it won't work out of the box.  For instance, the Capture.bundle file is not anticipated to live underneath another bundle, so the plugin accessing it would not work.

This is a first step toward updating to Cordova.  We can assess how to better package these resources after the update is complete.