Switching back to simply reloading the web view contents on 'reset'.
I think there's a bug in Cordova, with respect to tearing down and reloading the view controller.  No matter what I try, when I recreate the view controller (e.g. when the the login host is changed in settings), the splash screen stays on top forever.  I'll investigate the issue further, but in the meantime, I'm switching back to the way we've originally done it: just reload the start page in the existing web view.