Don't try to parse JSON for empty response.
Currently, REST responses are always attempted for JSON parsing of the response body.  Some responses, such as responses to updates, do not generate a response body.  Currently, there's a warning that's generated in these scenarios.  In reality, we should be bounds checking the response body, and not attempting JSON deserialization in that case.  This fixes that.