Normalizing native web view user agent with hybrid.
Apparently, making our user agent work properly in native web views also marginally messed up the OAuth web view; the layout/markup of the OAuth web view seems to be dependent on the device type (as sniffed through the user agent), and I didn't update the native side to tack on the original (iPad-ish) user agent to the string we create.  That was fun to track down.