Updating hybrid logout/reset functionality.
Switching to a more standard model of resetting an iOS application, destroying and recreating the root view controller when a user enters a situation where they have to log out or switch login hosts.  This removes the necessity of having to directly call functionality to load the start page in the Cordova view controller.