Library reorg
Changing a few things about the way the libraries are staged for building and consumption:

- Moving header search path configuration from "User header search paths" to "Header search paths".
- All library header files will now be deployed in Headers/[Project Name].
- Moved the copying of header files into a proper "Copy Header Files" build phase.
- Changed the build configuration to work with the 'install' action.

There's not a lot of net difference in the creation and consumption of the library archives; the new location of the header files is about it.  The native and hybrid templates have been updated to point to the new header file location.