While downloading the file on 12.0, file download is stuck on loading bar at 0%
https://gus.my.salesforce.com/apex/adm_bugdetail?id=a07B0000002QvWe&sfdc.override=1

I have investigated this issue (working in release-acadia).

The reason that progress is always 0, because _progress is never getting Initialized inside CSFSalesforceAction > setEnqueuedNetwork
And because of that, when adding observer to progress (progress is nil) is getting ignored.

And in order to do that , had to make progress readWrite inside CSFAction+Internal

Another way might be just call super inside CSFSalesforceAction > setEnqueuedNetwork
but I believe that is a bad idea (calling super inside custom setter), and it didn't even work.