Unregistering from notifications before starting actual logout
If we unregister after sending out the will logout notification, then the SFRestAPI instance for the user will have been torn down already.
The unregister call will cause a new SFRestAPI to be built for the user. That user object credentials will be taken away during logout but the SFRestAPI will stay.
If you log back in as the same user, you will get that SFRestAPI for your network calls. Because it holds a user without credentials, the first network call will cause the app to go back to login!