Fix for bug
Bug
a) cancel from SFAuthenticationSession page
b) you see host picker: pick custom domain (which was already selected)
c) you get stuck on splash screen

Why did it fail
The host list view controller was in the process of being dismissed when we get to the `    if ([self.authWindow.viewController presentedViewController]) {` causing the `dismissViewControllerAnimated` to be called.
However it would never complete, so the completion block and therefore `presentViewBlock()` was never being invoked.

Why it works now
We only want to dismiss the presented view controller if it's not in the process of being dismissed and if it's a SFLoginController.