Always doing running SFSDKWebViewStateManager's removeSession on the main thread
If you do a SFRestAPI send, and no credentials are available, SFUserAccountManager's loginWithCompletion is called, which calls SFUserAccountManager's authenticateWithcompletion => so you will be on whatever thread SFRestAPI send was invoked from
However if you do a SFRestAPI send and a refresh is needed and the refresh fails (e.g. refresh token was expired), then SFUserAccountManager's logoutUser is called, but that call happens is dispatched from the main thread.