Fixing bug in saveSoupEntryExternally causing externally stored files to be cut-off in some cases
When we upsert soup entries that are stored in external files, we use `NSJSONSerialization:writeJSONObject` to write the JSON to a file.
That method can fail if with certain unicode characters. See [corruption tester](https://github.com/wmathurin/CorruptionTester).
`saveSoupEntryExternally` should exit with an error, causing the `upsert` to revert.
Unfoturnately, we were looking at the return value `NSJSONSerialization:writeJSONObject` to decide if an error occurred.
But this method returns the number of bytes written so it could return a non zero value even though it fails to write the entire JSON.
As a result, the file will be cut-off, which will cause issues at query time.