Removing old crypto code (SFOAuthCrypto)
It was only kept in the SDK to allow for seamless upgrade.
Keeping it would only make sense if we believe there are apps that might jump from SDK 4.0 to SDK 8.0.

NB: we forgot to deprecate that class (and a few related public methods) in the last minor release.
However I don't think those were "true" public APIs.
Also we are getting flagged by DataTheorem for the old crypto calls.