Fix deadlock on [SFUserAccountManager sharedInstance]
When the `[SFUserAccountManager sharedInstance]` is created and sends the `SFUserAccountManagerDidFinishUserInitNotification`, if any observer attempts to access the `[SFUserAccountManager sharedInstance]` when the notification is posted, a deadlock is created.

This is because the `dispatch_once` block for posting the notification is still being executed when the `dispatch_once` token is accessed again during the call to access the `[SFUserAccountManager sharedInstance]`.

More information on why this is a problem can be found here:
https://stackoverflow.com/questions/19176219/why-am-i-getting-deadlock-with-dispatch-once

This PR fixes this bug, identified in https://github.com/forcedotcom/SalesforceMobileSDK-iOS/issues/3111