Key value stores not always cleared on logout
SalesforceSDKManager deletes the key value stores by listening for the logout notification and using handleUserWillLogout, this works when an app just uses SalesforceSDKManager. If an app uses SmartStoreSDKManager which is a subclass of SalesforceSDKManager instead, there's an issue because SmartStoreSDKManager uses the same handleUserWillLogout selector and only the one on SmartStoreSDKManager is triggered.

==> Solution using a different selector in SmartStoreSDKManager