Fix for failing testRestCalls
I debugged and verified the same cache was being used across NSURLSession but there seems to be NSURLSession session-level cache partitioning that causes the second round of calls to miss.
I decided to rewrite testRestCalls and use a test subclass of SFSDKEncryptedURLCache.
testRestCalls sends actual requests to the server twice. After the first set of calls, we make sure the cache was written to. After the second set of calls, we make sure the cache was read from.